Properly convert to ints

+++ FredBoat/src/main/java/fredboat/FredBoat.java
@@ -314,7 +314,7 @@ public abstract class FredBoat {
     }
 
     private static AtomicInteger biggestUserCountTotal = new AtomicInteger(-1); //of FredBoat total
-    public static long countAllUniqueUsers() {
+    public static int countAllUniqueUsers() {
         int result = JDAUtil.countAllUniqueUsers(shards, biggestUserCountTotal);
         //never shrink the user count (might happen due to not connected shards)
         biggestUserCountTotal.accumulateAndGet(result, Math::max);
